extract LeafletView and QrCode to /components
These two components can foreseeably be used across the application, and not just on any particular screen.
Thus, it seems logical to extract them to a common 'components' folder. In the future, we can also put other reusable components here to be used across the app.
