+++ www/js/components/QrCode.jsx
